requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

coverage
coveralls
dj-database-url
django
django-cors-headers
django-extensions
django-heroku
django-rest-swagger
djangorestframework
facebook-sdk
google-api-python-client
gunicorn
pipenv-to-requirements
psycopg2-binary
pyjwt
python-decouple
python-twitter
